fix(ai): address Tech Lead review — NULL safety + transaction + consolidate config resolution
- Add partial unique index for NULL channel (PostgreSQL NULL distinctness)
- Wrap updateAiConfig find+create/update in $transaction (TOCTOU race fix)
- Extract resolveAiConfig into shared src/lib/ai/resolve-config.ts
- Replace all inline config lookups in actions.ts and ai-agent.ts
- Remove null as unknown as ChannelType type hack
